Isolate umask callers to a separate process
